backend/course-service/requirements.txt
# Core dependencies
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
asyncio==3.4.3

# Database
SQLAlchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
uvloop==0.19.0; sys_platform != "win32"
pytest==7.4.3
pytest-asyncio==0.23.2 